requirements.txt
pytest>=7.0.0
pytest-json-report>=1.5.0
allure-pytest>=2.13.0
requests>=2.28.0
PyYAML>=6.0
openpyxl>=3.0.10
python-dotenv>=0.20.0
pymysql>=1.0.2
bcrypt>=4.0.0
